fix tests exiting and more tests
