+++ scripts/applyVotiumRewards.ts
@@ -1,3 +1,4 @@
+import { ethers } from "hardhat";
 import {
   votiumClaimRewards,
   votiumSellRewards,
@@ -6,8 +7,9 @@ import {
 const votiumStrategyAddress = "0xbbba116ef0525cd5ea9f4a9c1f628c3bfc343261";
 
 (async function main() {
-  const proofs = await votiumClaimRewards(votiumStrategyAddress);
-  await votiumSellRewards(votiumStrategyAddress, proofs);
+  const accounts = await ethers.getSigners();
+  const proofs = await votiumClaimRewards(accounts[0], votiumStrategyAddress);
+  await votiumSellRewards(accounts[0], votiumStrategyAddress, proofs);
 })()
   .then(() => process.exit(0))
   .catch((error) => {
