+++ scripts/applyVotiumRewardsHelpers.ts
@@ -1,8 +1,76 @@
 import { ethers } from "hardhat";
 import { VotiumErc20Strategy } from "../typechain-types";
 import axios from "axios";
-import { generate0xSwapData } from "./generateTestData";
 import { SignerWithAddress } from "@nomiclabs/hardhat-ethers/signers";
+import { wethAbi } from "../test/abis/wethAbi";
+import { BigNumber } from "ethers";
+
+export const generate0xSwapData = async (
+  tokenAddresses: string[],
+  tokenAmounts: string[]
+) => {
+  const accounts = await ethers.getSigners();
+
+  const swapsData = [];
+  // swap reward tokens for eth
+  for (let i = 0; i < tokenAddresses.length; i++) {
+    const sellToken = tokenAddresses[i];
+    const buyToken = "0xEeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE";
+
+    // we use weth abi because we sometimes need to call withdraw on weth but its otherwise an erc20 abi
+    const tokenContract = new ethers.Contract(
+      tokenAddresses[i],
+      wethAbi,
+      accounts[0]
+    );
+
+    const sellAmount = BigNumber.from(tokenAmounts[i]);
+
+    // special case unwrap weth
+    if (
+      sellToken.toLowerCase() ===
+      "0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2".toLowerCase()
+    ) {
+      const data = await tokenContract.populateTransaction.withdraw(sellAmount);
+      const newData = {
+        sellToken,
+        spender: tokenContract.address,
+        swapTarget: tokenContract.address,
+        swapCallData: data.data,
+      };
+      swapsData.push(newData);
+    } else {
+      let result;
+      // TODO do we want slippage protection or does it not matter and we just dump all the tokens anyway?
+      try {
+        result = await axios.get(
+          `https://api.0x.org/swap/v1/quote?buyToken=${buyToken}&sellToken=${sellToken}&sellAmount=${sellAmount}`,
+          {
+            headers: {
+              "0x-api-key":
+                process.env.API_KEY_0X ||
+                "35aa607c-1e98-4404-ad87-4bed10a538ae",
+            },
+          }
+        );
+
+        const newData = {
+          sellToken,
+          spender: result.data.allowanceTarget,
+          swapTarget: result.data.to,
+          swapCallData: result.data.data,
+        };
+
+        swapsData.push(newData);
+      } catch (e) {
+        console.log("0x doesnt support", i, sellToken, buyToken, sellAmount, e);
+      }
+    }
+    // prevent 429s
+    await new Promise((resolve) => setTimeout(resolve, 500));
+  }
+  return swapsData;
+};
 
 // Claims all rewards using public votium merkle proofs
 // or pass in proofs to override
