+++ test/strategies/VotiumErc20/VotiumErc20Strategy.test.ts
@@ -254,12 +254,12 @@ describe("Test VotiumErc20Strategy", async function () {
     const priceBeforeRewards = await votiumStrategy.price();
 
     await votiumClaimRewards(
-      accounts[0],
+      rewarderAccount,
       votiumStrategy.address,
       testData.claimProofs
     );
     await votiumSellRewards(
-      accounts[0],
+      rewarderAccount,
       votiumStrategy.address,
       [],
       testData.swapsData
